# apps/bs-network-backend/src/main/java/com/eustache/book_network/book/BookResponse.java
package com.eustache.book_network.book;

import lombok.*;

@Getter
@Setter
@AllArgsConstructor
@NoArgsConstructor
@Builder
public class BookResponse {
  private Integer id;
  private String title;
  private String authorName;
  private String isbn;
  private String synopsis;
  private String owner;
  private byte[] cover;
  private double rate;
  private boolean archived;
  private boolean shareable;
}