# apps/bs-network-backend/src/main/java/com/eustache/book_network/book/BorrowedBookResponse.java
package com.eustache.book_network.book;

import lombok.*;

@Getter
@Setter
@AllArgsConstructor
@NoArgsConstructor
@Builder
public class BorrowedBookResponse {
  private Integer id;
  private String title;
  private String authorName;
  private String isbn;
  private double rate;
  private boolean returned;
  private boolean returnApproved;
}