# apps/bs-network-backend/src/main/java/com/eustache/book_network/feedback/FeedbackController.java
package com.eustache.book_network.feedback;

import com.eustache.book_network.common.PageResponse;
import io.swagger.v3.oas.annotations.tags.Tag;
import jakarta.validation.Valid;
import lombok.RequiredArgsConstructor;
import org.springframework.http.ResponseEntity;
import org.springframework.security.core.Authentication;
import org.springframework.web.bind.annotation.*;

@RestController
@RequestMapping("feedbacks")
@RequiredArgsConstructor
@Tag(name = "Feedback")
public class FeedbackController {
  private final FeedbackService service;

  @PostMapping
  public ResponseEntity<Integer> saveFeedback(
    @Valid @RequestBody FeedbackRequest request,
    Authentication connectedUser
  ) {
    return ResponseEntity.ok(service.save(request, connectedUser));
  }

  @GetMapping("/book/{book-id}")
  public ResponseEntity<PageResponse<FeedbackResponse>> findAllFeedbacksByBook(
    @PathVariable("book-id") Integer bookId,
    @RequestParam(name = "page", defaultValue = "0", required = false) int page,
    @RequestParam(name = "size", defaultValue = "10", required = false) int size,
    Authentication connectedUser
  ) {
    return ResponseEntity.ok(service.findAllFeedbacksByBook(bookId, page, size, connectedUser));
  }
}