# apps/bs-network-backend/src/main/java/com/eustache/book_network/feedback/FeedbackResponse.java
package com.eustache.book_network.feedback;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;

@Getter
@Setter
@AllArgsConstructor
@NoArgsConstructor
@Builder
public class FeedbackResponse {
  private Double note;
  private String comment;
  private boolean ownFeedback;
}